Selenium/requirements.txt
selenium
english-words
pillow
numpy